Code style updates. Moved the new progress bar into a Controls folder and namespace.
