+++ src/Kuriimu2_WinForms/Kuriimu2.cs
@@ -1,23 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Linq;
-using System.Text;
 using System.IO;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using Kuriimu2_WinForms.FormatForms;
-using Kontract.Interfaces.Text;
-using Kontract.Interfaces.Image;
-using Kontract.Interfaces.Archive;
-using Kontract.Interfaces.Common;
-using System.Text.RegularExpressions;
-using Kontract.Interfaces.FileSystem;
 using Kontract.FileSystem;
-using Kuriimu2_WinForms.Interfaces;
+using Kontract.Interfaces.Archive;
+using Kontract.Interfaces.Image;
+using Kontract.Interfaces.Text;
 using Kore;
+using Kuriimu2_WinForms.FormatForms;
+using Kuriimu2_WinForms.Interfaces;
 using Kuriimu2_WinForms.Properties;
 
 namespace Kuriimu2_WinForms
