+++ src/Kuriimu2_WinForms/Kuriimu2_WinForms.csproj
@@ -50,7 +50,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="InfoProgressBar.cs">
+    <Compile Include="Controls\InfoProgressBar.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="InputBox.cs">
